Medical-AI-Assistant/backend/requirements.txt
# Medical AI Assistant Backend - Python Dependencies
# Generated for use with pip install -r requirements.txt
# For production deployments

# =============================================================================
# Core Framework
# =============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# =============================================================================
# Database
# =============================================================================
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# =============================================================================
# Authentication & Security
# =============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
python-multipart==0.0.6

# =============================================================================
# WebSocket & Async
# =============================================================================
websockets==12.0
redis==5.0.1
aioredis==2.0.1

# =============================================================================
# LangChain & Agent Framework
# =============================================================================
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10
langsmith==0.0.77

# =============================================================================
# Vector Store & RAG
# =============================================================================
chromadb==0.4.22
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# =============================================================================
# ML & Model Inference
# =============================================================================
transformers==4.37.0
torch==2.1.2
accelerate==0.26.1
bitsandbytes==0.42.0
peft==0.8.2
deepspeed==0.13.1
safetensors==0.4.2
huggingface-hub==0.20.2

# =============================================================================
# Data Processing
# =============================================================================
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.26.0
tenacity==8.2.3
pyyaml==6.0.1

# =============================================================================
# Monitoring & Observability
# =============================================================================
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# =============================================================================
# Development & Testing (commented out for production)
# =============================================================================
# pytest==7.4.4
# pytest-asyncio==0.23.3
# pytest-cov==4.1.0
# pytest-mock==3.12.0
# black==24.1.1
# flake8==7.0.0
# mypy==1.8.0
# isort==5.13.2
# pre-commit==3.6.0
